Add final edit functionality.
With this commit, you should be able to successfully
edit a journal entry.  The old journal
entry should be completely removed from the DB
and a new journal entry put in its place with the edits.
On delete cascade is not working with the PDO
object so linking table entries are explicitly deleted.